pyproject.toml
[tool.poetry]
name = "saas-group"
version = "0.1.0"
description = ""
authors = ["Damilola Adeyemi <adeyemidamilola3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
aerich = "^0.7.2"
pydantic-settings = "^2.2.1"
tenacity = "^8.2.3"
httpx = "^0.27.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-env = "^1.1.3"
asgi-lifespan = "^2.1.0"
pytest-mock = "^3.14.0"
gunicorn = "^21.2.0"

[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
asyncio_mode = "auto"
addopts = "-ra -q"
testpaths = [
    "**/**/tests",
]
env = [
    "ENV = test",
]
